+++ src/main/java/gr/katsip/synefo/storm/operators/synefo_comp_ops/Select.java
@@ -304,8 +304,9 @@ public class Select implements Serializable, AbstractStatOperator  {
 		int throughput = 0;
 		int sel = 0;
 		if(stats != null) {
-			String tuple = 	stats.getCpuLoad() + "," + stats.getMemory() + "," + stats.getWindowLatency() + "," + 
-					stats.getWindowThroughput() + "," + stats.getSelectivity() + "," + 
+			String tuple = 	(float) stats.getCpuLoad() + "," + (float) stats.getMemory() + "," + 
+					(int) stats.getWindowLatency() + "," + (int) stats.getWindowThroughput() + "," + 
+					(float) stats.getSelectivity() + "," + 
 					encryptionData.get("pln") + "," +
 					encryptionData.get("DET") + "," +
 					encryptionData.get("RND") + "," +
